+++ CHANGELOG.md
@@ -30,6 +30,9 @@ Test/interim releases are not documented here.
 - Transpiler - add default options to output finlang version, date/time, and MD5 filepath.
     - Also outputs source file path.
 - Transpiler - Support `SimOnly.run(() => { /* code */ })` for running code in simulation mode only.
+- Transpiler - Support C# binary/hex literals with underscore separators
+    - `0b1010_1110` --> `0b10101110`
+    - `0x1234_5678` --> `0x12345678`
 
 ---